--- browser/src/main/java/nolambda/uibook/browser/form/ComposeInputCreator.kt
@@ -1,20 +1,0 @@
-package nolambda.uibook.browser.form
-
-import androidx.compose.runtime.Composable
-import nolambda.uibook.annotations.FunctionParameter
-
-/**
- * Input creator is a factory interface to create the input in UI Book
- * This input add interactivity to the book so user can understand the book better
- *
- * @see [DefaultInputCreator]
- */
-interface ComposeInputCreator {
-
-    @Composable
-    fun CreateInput(
-        parameter: FunctionParameter,
-        defaultState: Any,
-        setViewState: (Any) -> Unit,
-    )
-}
