+++ browser/src/main/java/nolambda/uibook/components/bookform/InputData.kt
@@ -1,9 +1,9 @@
 package nolambda.uibook.components.bookform
 
-import nolambda.uibook.browser.form.ComposeInputCreator
+import nolambda.uibook.browser.form.InputCreator
 
 class InputData(
     val viewState: Array<Any>,
-    val inputCreator: ComposeInputCreator,
+    val inputCreator: InputCreator,
     val setViewState: (Int, Any) -> Unit
 )
