+++ sample/src/main/java/nolambda/uibook/samples/ButtonInput.kt
@@ -1,28 +1,27 @@
 package nolambda.uibook.samples
 
-import android.view.LayoutInflater
-import android.view.View
-import android.widget.Button
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material.Button
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import nolambda.uibook.annotations.FunctionParameter
-import nolambda.uibook.browser.databinding.ViewFormBinding
 import nolambda.uibook.browser.form.InputCreator
 import java.util.*
 
 class ButtonInput : InputCreator {
 
-    override fun createInput(
-        inflater: LayoutInflater,
-        parent: ViewFormBinding,
+    @Composable
+    override fun CreateInput(
         parameter: FunctionParameter,
         defaultState: Any,
         setViewState: (Any) -> Unit
-    ): View {
-        val context = parent.root.context
-        return Button(context).apply {
-            text = "Random UUID"
-            setOnClickListener {
-                setViewState(UUID.randomUUID().toString())
-            }
+    ) {
+        Button(
+            onClick = { setViewState(UUID.randomUUID().toString()) },
+            modifier = Modifier.fillMaxWidth()
+        ) {
+            Text(text = "Random UUID")
         }
     }
 }
